dialog on remove and less css

+++ src/app/components/todo-list/todo-list.component.ts
@@ -1,14 +1,24 @@
-import { Component, OnInit, Inject, Output, EventEmitter } from "@angular/core";
+import {
+  Component,
+  OnInit,
+  Inject,
+  Output,
+  EventEmitter,
+  ViewChild
+} from "@angular/core";
 import { TodoModel } from "../../models/todo.model";
 import { TodoService } from "../../services/todo.service";
+import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import { DialogComponent } from "../dialog/dialog.component";
 
 @Component({
   selector: "todo-list",
   templateUrl: "./todo-list.component.html",
-  styleUrls: ["./todo-list.component.css".toString()]
+  styles: [require("./todo-list.component.less").toString()]
 })
 export class TodoListComponent implements OnInit {
   private todoes: Array<TodoModel>;
+  @ViewChild("removeDialog") removeDialog: DialogComponent;
   @Output() editTodo = new EventEmitter<TodoModel>();
 
   constructor(@Inject(TodoService) private todoService: TodoService) {}
@@ -35,12 +45,18 @@ export class TodoListComponent implements OnInit {
     this.todoService.toggleTodoCompleted(id);
   }
 
-  private removeTodo(id: number): void {
-    this.todoService.removeTodo(id);
+  private removeTodo(id: number, content: string): void {
+    this.removeDialog.open().then(
+      response => {
+        if (response) {
+          this.todoService.removeTodo(id);
+        }
+      },
+      dismiss => {}
+    );
   }
 
   private updateTodo(todo: TodoModel): void {
     this.editTodo.emit(todo);
   }
-
 }
